[jpa] chap11. jpashop 예제 시작 전 셋팅
