+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'jpa-ex'
+rootProject.name = 'jpa-jpashop'
