+++ src/main/java/jpabook/JpashopApplication.java
@@ -1,0 +1,12 @@
+package jpabook;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class JpashopApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(JpashopApplication.class, args);
+    }
+}
