+++ src/main/java/jpabook/jpashop/domain/Address.java
@@ -1,4 +1,4 @@
-package jpabook.ex;
+package jpabook.jpashop.domain;
 
 import javax.persistence.Embeddable;
