+++ src/main/java/jpabook/jpashop/domain/Delivery.java
@@ -1,26 +1,27 @@
-package jpabook.ex;
+package jpabook.jpashop.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import java.util.ArrayList;
-import java.util.List;
+import javax.persistence.OneToOne;
 
 /**
- * Created by sykim on 2016. 3. 20..
+ * Created by sykim on 2016. 3. 23..
  */
 @Entity
-public class Member extends BaseEntity {
+public class Delivery {
 
     @Id
     @GeneratedValue
-    @Column(name = "MEMBER_ID")
+    @Column(name = "DELIVERY_ID")
     private Long id;
 
-    private String name;
+    @OneToOne(mappedBy = "delivery")
+    private Order order;
 
     //private String city;
     //private String street;
@@ -29,11 +30,11 @@ public class Member extends BaseEntity {
     @Embedded
     private Address address;
 
-    @OneToMany(mappedBy = "member")
-    private List<Order> orders = new ArrayList<Order>();
+    @Enumerated(EnumType.STRING)
+    private DeliveryStatus status;
 
+    //Getter, Setter
 
-    // Getter, Setter
     public Long getId() {
         return id;
     }
@@ -42,12 +43,12 @@ public class Member extends BaseEntity {
         this.id = id;
     }
 
-    public String getName() {
-        return name;
+    public Order getOrder() {
+        return order;
     }
 
-    public void setName(String name) {
-        this.name = name;
+    public void setOrder(Order order) {
+        this.order = order;
     }
 
     public Address getAddress() {
@@ -58,11 +59,11 @@ public class Member extends BaseEntity {
         this.address = address;
     }
 
-    public List<Order> getOrders() {
-        return orders;
+    public DeliveryStatus getStatus() {
+        return status;
     }
 
-    public void setOrders(List<Order> orders) {
-        this.orders = orders;
+    public void setStatus(DeliveryStatus status) {
+        this.status = status;
     }
 }
