+++ src/main/java/jpabook/jpashop/domain/Member.java
@@ -1,27 +1,26 @@
-package jpabook.ex;
+package jpabook.jpashop.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-import javax.persistence.OneToOne;
+import javax.persistence.OneToMany;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
- * Created by sykim on 2016. 3. 23..
+ * Created by sykim on 2016. 3. 20..
  */
 @Entity
-public class Delivery {
+public class Member extends BaseEntity {
 
     @Id
     @GeneratedValue
-    @Column(name = "DELIVERY_ID")
+    @Column(name = "MEMBER_ID")
     private Long id;
 
-    @OneToOne(mappedBy = "delivery")
-    private Order order;
+    private String name;
 
     //private String city;
     //private String street;
@@ -30,11 +29,11 @@ public class Delivery {
     @Embedded
     private Address address;
 
-    @Enumerated(EnumType.STRING)
-    private DeliveryStatus status;
+    @OneToMany(mappedBy = "member")
+    private List<Order> orders = new ArrayList<Order>();
 
-    //Getter, Setter
 
+    // Getter, Setter
     public Long getId() {
         return id;
     }
@@ -43,12 +42,12 @@ public class Delivery {
         this.id = id;
     }
 
-    public Order getOrder() {
-        return order;
+    public String getName() {
+        return name;
     }
 
-    public void setOrder(Order order) {
-        this.order = order;
+    public void setName(String name) {
+        this.name = name;
     }
 
     public Address getAddress() {
@@ -59,11 +58,11 @@ public class Delivery {
         this.address = address;
     }
 
-    public DeliveryStatus getStatus() {
-        return status;
+    public List<Order> getOrders() {
+        return orders;
     }
 
-    public void setStatus(DeliveryStatus status) {
-        this.status = status;
+    public void setOrders(List<Order> orders) {
+        this.orders = orders;
     }
 }
