+++ src/main/java/jpabook/jpashop/domain/Order.java
@@ -1,4 +1,4 @@
-package jpabook.ex;
+package jpabook.jpashop.domain;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
