+++ src/main/java/jpabook/jpashop/domain/item/Movie.java
@@ -1,4 +1,4 @@
-package jpabook.ex.item;
+package jpabook.jpashop.domain.item;
 
 import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
