+++ src/main/resources/META-INF/persistence.xml
@@ -1,33 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <persistence xmlns="http://xmlns.jcp.org/xml/ns/persistence" version="2.1">
-
-    <persistence-unit name="jpa-book">
-
-        <class>jpabook.ex.Member</class>
-        <class>jpabook.ex.Order</class>
-        <class>jpabook.ex.OrderItem</class>
-        <class>jpabook.ex.item.Item</class>
-        <class>jpabook.ex.Delivery</class>
-        <class>jpabook.ex.Category</class>
-
-        <properties>
-
-            <!-- 필수 속성 -->
-            <property name="javax.persistence.jdbc.driver" value="org.h2.Driver"/>
-            <property name="javax.persistence.jdbc.user" value="sa"/>
-            <property name="javax.persistence.jdbc.password" value=""/>
-            <property name="javax.persistence.jdbc.url" value="jdbc:h2:tcp://localhost/~/test"/>
-            <property name="hibernate.dialect" value="org.hibernate.dialect.H2Dialect" />
-
-            <!-- 옵션 -->
-            <property name="hibernate.show_sql" value="true" />
-            <property name="hibernate.format_sql" value="true" />
-            <property name="hibernate.use_sql_comments" value="true" />
-            <property name="hibernate.id.new_generator_mappings" value="true" />
-
-            <!-- 데이터베이스 스키마 자동 생성 -->
-            <property name="hibernate.hbm2ddl.auto" value="create" />
-        </properties>
-    </persistence-unit>
-
+    <persistence-unit name="mock" />
 </persistence>
